WT-16261 Refactor binary decode log parsing. Support parsing multiple byte dumps.
